# documentation/openapi.json
{
  "components": {
    "schemas": {
      "Company": {
        "type": "object",
        "properties": {
          "voucherList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Voucher"
            }
          },
          "city": {
            "type": "string"
          },
          "twitterUrl": {
            "format": "url",
            "type": "string"
          },
          "postalCode": {
            "format": "int32",
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "uuid": {
            "format": "uuid",
            "type": "string"
          },
          "needHelpBecause": {
            "type": "string"
          },
          "logoUrl": {
            "format": "url",
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "facebookUrl": {
            "format": "url",
            "type": "string"
          },
          "wallpaperUrl": {
            "format": "url",
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "paypalAddress": {
            "type": "string"
          },
          "instagramUrl": {
            "format": "url",
            "type": "string"
          },
          "sector": {
            "$ref": "#/components/schemas/Sector"
          },
          "imageList": {
            "type": "array",
            "items": {
              "format": "url",
              "type": "string"
            }
          },
          "homepage": {
            "format": "url",
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "voucherList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Voucher"
            }
          },
          "companyList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Company"
            }
          },
          "uuid": {
            "format": "uuid",
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "Sector": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "uuid": {
            "format": "uuid",
            "type": "string"
          }
        }
      },
      "Voucher": {
        "type": "object",
        "properties": {
          "voucherId": {
            "type": "string"
          },
          "company": {
            "$ref": "#/components/schemas/Company"
          },
          "defaultValueInEurCt": {
            "format": "int32",
            "type": "integer"
          },
          "used": {
            "type": "boolean"
          },
          "uuid": {
            "format": "uuid",
            "type": "string"
          },
          "valueInEurCt": {
            "format": "int32",
            "type": "integer"
          }
        }
      }
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
  "openapi": "3.0.1",
  "paths": {
    "/company": {
      "post": {
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/Company"
              }
            }
          }
        },
        "operationId": "createCompany",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          }
        },
        "tags": [
          "Companies"
        ]
      },
      "get": {
        "operationId": "getCompanies",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Companies"
        ]
      }
    },
    "/voucher": {
      "post": {
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/Voucher"
              }
            }
          }
        },
        "operationId": "createVoucher",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voucher"
                }
              }
            }
          }
        },
        "tags": [
          "Vouchers"
        ]
      },
      "get": {
        "operationId": "getVouchers",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Vouchers"
        ]
      }
    },
    "/company/{uuid}": {
      "get": {
        "operationId": "getCompanyById",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "uuid",
            "required": true
          }
        ],
        "tags": [
          "Companies"
        ]
      },
      "delete": {
        "operationId": "deleteCompany",
        "responses": {
          "200": {
            "description": "default response"
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "uuid",
            "required": true
          }
        ],
        "tags": [
          "Companies"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/Company"
              }
            }
          }
        },
        "operationId": "updateCompany",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "uuid",
            "required": true
          }
        ],
        "tags": [
          "Companies"
        ]
      }
    },
    "/voucher/{uuid}/current_amount": {
      "put": {
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "format": "int32",
                "type": "integer"
              }
            }
          }
        },
        "operationId": "updateVoucherAmount",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voucher"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "uuid",
            "required": true
          }
        ],
        "tags": [
          "Vouchers"
        ]
      }
    },
    "/sector": {
      "get": {
        "operationId": "getSectors",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Sectors"
        ]
      }
    },
    "/voucher/{uuid}": {
      "get": {
        "operationId": "getVoucherById",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voucher"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "uuid",
            "required": true
          }
        ],
        "tags": [
          "Vouchers"
        ]
      },
      "delete": {
        "operationId": "deleteVoucher",
        "responses": {
          "200": {
            "description": "default response"
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "uuid",
            "required": true
          }
        ],
        "tags": [
          "Vouchers"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/Voucher"
              }
            }
          }
        },
        "operationId": "updateVoucher",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voucher"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "uuid",
            "required": true
          }
        ],
        "tags": [
          "Vouchers"
        ]
      }
    },
    "/user": {
      "post": {
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "operationId": "createUser",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/user/{uuid}": {
      "get": {
        "operationId": "getUserById",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "uuid",
            "required": true
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "deleteUser",
        "responses": {
          "200": {
            "description": "default response"
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "uuid",
            "required": true
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "operationId": "updateUser",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "uuid",
            "required": true
          }
        ],
        "tags": [
          "Users"
        ]
      }
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  }
}